Get smart about how modifier key fallbacks work
Allow some lazy typing, but make sure the amount of modifier keys
pressed matches the intended sequence. For example: pressing `C-a C-r`
should not match a binding for `C a r` because the input sequence has
two Control modifiers, but the required sequence only has one.
Pressing `C-a` should match the binding for `C a` because there are
the correct number of control keys in the sequence.