Get much smarter about the way modifier keys are handled
That's what I get for trying to be smart while I was damn new falling
asleep last night. This is a much better way of handling modifier
keys that should hopefully feel as natural as possible. For example,
if there's a keybinding that asks for `Control+r Control+r`, the user
can hold down Control and hit 'r' twice, which makes ergonomic sense.

The previous method of inserting bindings into the tree not only made
this impossible, it was also way more complicated, and the dispatcher
code for then dealing with the weird storage was also complicated.
Instead of caring about intermediate modifier keypresses, which we
were trying to do before, when they don't match a node we just remove
it from the sequence and ignore it.